+++ src/App.js
@@ -30,6 +30,7 @@ function App() {
     if (user === '') {
       navigate('/');
     }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [sesion]);
 
   return (
