+++ src/App_original.js
@@ -20,21 +20,21 @@ import { Routes, Route } from 'react-router-dom';
 
 function App() {
   //para la sesion
-  const navigate = useNavigate();
-  const [sesion, setSesion] = useState('');
+  // const navigate = useNavigate();
+  // const [sesion, setSesion] = useState('');
 
   //Todo esto podría ir a la store global:
   // const [canchas, setCanchas] = useState([]);
-  const [reservas, setReservas] = useState([]);
+  // const [reservas, setReservas] = useState([]);
   // const [pagos, setPagos] = useState([]);
 
   // const [actCanchas, setActCanchas] = useState(false);
-  const [actReservas, setActReservas] = useState(false);
+  // const [actReservas, setActReservas] = useState(false);
   // const [actPagos, setActPagos] = useState(false);
   /*  */
   /* Loaders */
   const [activedLoader, setActivedLoader] = useState(false);
-  const [reservasLoader, setReservasLoader] = useState(false);
+  // const [reservasLoader, setReservasLoader] = useState(false);
   // const [alumnosLoader, setAlumnosLoader] = useState(false);
   // const [profesoresLoader, setProfesoresLoader] = useState(false);
 
@@ -78,17 +78,17 @@ function App() {
   // }, [actCanchas]);
 
   //Con las reservas
-  useEffect(() => {
-    const requestOptions = {
-      method: 'GET',
-    };
-    fetch(`${URL_BASE}reservas`, requestOptions)
-      .then(setReservasLoader(true))
-      .then((response) => response.json())
-      .then((data) => setReservas(data.detail))
-      .then((response) => setReservasLoader((v) => false));
-    /* Desactivar Spinner */
-  }, [actReservas]);
+  // useEffect(() => {
+  //   const requestOptions = {
+  //     method: 'GET',
+  //   };
+  //   fetch(`${URL_BASE}reservas`, requestOptions)
+  //     .then(setReservasLoader(true))
+  //     .then((response) => response.json())
+  //     .then((data) => setReservas(data.detail))
+  //     .then((response) => setReservasLoader((v) => false));
+  //   /* Desactivar Spinner */
+  // }, [actReservas]);
 
   //get de los alumnos
   // useEffect(() => {
@@ -104,17 +104,17 @@ function App() {
   // }, [actAlumnos]);
 
   //get Profesores
-  useEffect(() => {
-    const requestOptions = {
-      method: 'GET',
-    };
-    fetch(`${URL_BASE}profesores`, requestOptions)
-      .then((response) => response.json())
-      .then((data) => setProfesores(ordenarPorNombre(data)))
-      .then((response) => setProfesoresLoader((v) => false));
+  // useEffect(() => {
+  //   const requestOptions = {
+  //     method: 'GET',
+  //   };
+  //   fetch(`${URL_BASE}profesores`, requestOptions)
+  //     .then((response) => response.json())
+  //     .then((data) => setProfesores(ordenarPorNombre(data)))
+  //     .then((response) => setProfesoresLoader((v) => false));
 
-    /* Desactivar spinner */
-  }, [actProfesores]);
+  //   /* Desactivar spinner */
+  // }, [actProfesores]);
 
   //fet pagos
   // useEffect(() => {
@@ -128,101 +128,8 @@ function App() {
   // }, [actPagos]);
 
   return (
-    <>
-      <div className="App">
-        <header className="App-header"></header>
-        <Routes>
-          <Route path="/" element={<Login setSesion={setSesion} />}></Route>
-          <Route
-            path="/inicio"
-            element={<Home setSesion={setSesion} />}
-          ></Route>
-          <Route
-            path="/reservas"
-            element={
-              <HomeV
-                canchas={canchas}
-                reservas={reservas}
-                reservasLoader={reservasLoader}
-                setSesion={setSesion}
-                alumnos={alumnos}
-                profesores={profesores}
-                setActReservas={setActReservas}
-              />
-            }
-          ></Route>
-          <Route
-            path="/canchas"
-            element={
-              <Canchas
-                canchas={canchas}
-                setActCanchas={setActCanchas}
-                activedLoader={activedLoader}
-                setActivedLoader={setActivedLoader}
-                setSesion={setSesion}
-              />
-            }
-          ></Route>
-          <Route
-            path="/alumnos"
-            element={
-              <Alumnos
-                actAlumnos={actAlumnos}
-                setActAlumnos={setActAlumnos}
-                alumnos={alumnos}
-                setAlumnos={setAlumnos}
-                setAlumnosLoader={setAlumnosLoader}
-                alumnosLoader={alumnosLoader}
-                setSesion={setSesion}
-              />
-            }
-          ></Route>
-          <Route
-            path="/profesores"
-            element={
-              <Profesores
-                actProfesores={actProfesores}
-                setActProfesores={setActProfesores}
-                profesores={profesores}
-                setProfesores={setProfesores}
-                setProfesoresLoader={setProfesoresLoader}
-                profesoresLoader={profesoresLoader}
-                setSesion={setSesion}
-              />
-            }
-          ></Route>
-          <Route
-            path="/pagos"
-            element={
-              <Pagos
-                pagos={pagos}
-                actPagos={actPagos}
-                setActPagos={setActPagos}
-                setPagos={setPagos}
-                alumnos={alumnos}
-                setSesion={setSesion}
-              />
-            }
-          />
-          <Route
-            path="/nuevaReserva"
-            element={
-              <Reservas
-                canchas={canchas}
-                reservas={reservas}
-                setActReservas={setActReservas}
-                setReservasLoader={setReservasLoader}
-                setSesion={setSesion}
-                profesores={profesores}
-                alumnos={alumnos}
-                setProfesores={setProfesores}
-              />
-            }
-          ></Route>
-        </Routes>
-      </div>
-    </>
-  );
+    <></>
+  )
 }
 
 export default App;
